+++ src/room/room.component.tsx
@@ -1,23 +1,15 @@
 import { Header } from "../page-layout/header.component";
 import { Deck } from "../deck/deck.component";
-import { useEffect } from "react";
 import { VoteResult } from "../voting/vote-result.component";
-import { usePlayer } from "../player/player.context";
 import { useDealer } from "../dealer/dealer.context";
 import { Flex } from "antd";
 import { AddIssue } from "../issue/components/add-issue.component";
 import { IssueList } from "../issue/components/issue-list.component";
-import { useAppState, useUpdateAppState } from "../app/app.state";
+import { useAppState } from "../app/app.state";
 
 export function Room() {
-  const player = usePlayer()!;
   const dealer = useDealer();
   const appState = useAppState();
-  const updateAppState = useUpdateAppState();
-
-  useEffect(() => {
-    player.onStateChanged(updateAppState).enableHeartBeat();
-  }, [player, updateAppState]);
 
   return (
     <>